+++ src/connection/database/login_pane.jsx
@@ -3,7 +3,7 @@ import EmailPane from '../../field/email/email_pane';
 import UsernamePane from '../../field/username/username_pane';
 import PasswordPane from '../../field/password/password_pane';
 import { showResetPasswordActivity } from './actions';
-import { authWithUsername, hasScreen, forgotPasswordLink } from './index';
+import { hasScreen, forgotPasswordLink } from './index';
 import * as l from '../../core/index';
 
 export default class LoginPane extends React.Component {
@@ -35,6 +35,7 @@ export default class LoginPane extends React.Component {
       ? <EmailPane
           i18n={i18n}
           lock={lock}
+          forceInvalidVisibility={!showPassword}
           placeholder={emailInputPlaceholder}
         />
       : <UsernamePane