+++ src/engine/classic/sign_up_pane.jsx
@@ -2,12 +2,12 @@ import React from 'react';
 import EmailPane from '../../field/email/email_pane';
 import PasswordPane from '../../field/password/password_pane';
 import UsernamePane from '../../field/username/username_pane';
+import CustomInput from '../../field/custom_input';
 import {
   additionalSignUpFields,
   databaseConnectionRequiresUsername,
   passwordStrengthPolicy,
 } from '../../connection/database/index';
-import CustomInput from '../../field/custom_input';
 
 export default class SignUpPane extends React.Component {